refactor: move to manifest package
