+++ packages/fx-core/src/component/driver/teamsApp/configure.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-import { FxError, Result, err, ok, Platform } from "@microsoft/teamsfx-api";
+import { FxError, Result, err, ok, ManifestUtil } from "@microsoft/teamsfx-api";
 import fs from "fs-extra";
 import { hooks } from "@feathersjs/hooks/lib";
 import isUUID from "validator/lib/isUUID";
@@ -18,7 +18,6 @@ import { manifestUtils } from "./utils/ManifestUtils";
 import { AppStudioError } from "./errors";
 import { AppStudioScopes } from "../../../common/tools";
 import { getLocalizedString } from "../../../common/localizeUtils";
-import { TelemetryProperty } from "../../../common/telemetry";
 import { Service } from "typedi";
 import { getAbsolutePath } from "../../utils/common";
 import { FileNotFoundError, InvalidActionInputError } from "../../../error/common";
@@ -98,7 +97,7 @@ export class ConfigureTeamsAppDriver implements StepDriver {
       return err(manifest.error);
     }
 
-    const manifestTelemetries = manifestUtils.parseCommonTelemetryProperties(manifest.value);
+    const manifestTelemetries = ManifestUtil.parseCommonTelemetryProperties(manifest.value);
     merge(context.telemetryProperties, manifestTelemetries);
 
     // Fail if Teams app not exists, as this action only update the Teams app, not create
