+++ packages/manifest/src/constants.ts
@@ -1,0 +1,4 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+
+export const SharePointAppId = "00000003-0000-0ff1-ce00-000000000000";
